Main Code
The Actual code of the moving average purchasing simulator. The
algorithm can be tweaked quite easily inside the if else statement (see
comments).

The algorithm is profitable at values of price>1.005*trend. I have kept
price<trend for SELL.

Next step is to make it actually work together with Kraken or Bitpay
(probably Kraken as the system is based on Kraken API).